2023/10/19 : 19:31 add image and modify dto, repository

+++ src/main/java/repository/UserRepository.java
@@ -1,7 +1,9 @@
 package repository;
 
-import domain.User;
+import dto.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
+
+
 }
